refactor(core): repoint catalog consumers to SSOT and delete legacy wrappers
