+++ backend/api/src/controllers/admin/catalog/adminModelController.ts
@@ -4,16 +4,15 @@ import { sendSuccessResponse } from '../../../utils/respond';
 import { handlePaginatedContent } from '../../../utils/contentHandler';
 import mongoose from 'mongoose';
 import { CATALOG_APPROVAL_STATUS } from "@esparex/contracts";
-import { CatalogModel, findCategoryBySlugForCatalog, getActiveBrandIds, checkBrandInCategories, checkModelDependencies, findModelByFilter, findModelBySlug } from '@esparex/core/services/catalog/CatalogBrandModelService';
-import { validateBrandIsActive } from '@esparex/core/services/catalog/CatalogValidationService';
-import CatalogOrchestrator from '@esparex/core/services/catalog/CatalogOrchestrator';
+import { CatalogModel, findCategoryBySlugForCatalog, getActiveBrandIds, checkBrandInCategories, checkModelDependencies, findModelByFilter, findModelBySlug } from '@esparex/core/domains/catalog/application/services/CatalogBrandModelService';
+import { validateBrandIsActive } from '@esparex/core/domains/catalog/application/services/CatalogValidationService';
+import CatalogOrchestrator from '@esparex/core/domains/catalog/application/services/CatalogOrchestrator';
 import { invalidateItemCatalogCache } from './shared';
 import { sendCatalogError, QueryRecord, ACTIVE_CATEGORY_QUERY, validateActiveCategories, getActiveCategoryIds, handleCatalogCreate, handleCatalogUpdate, handleCatalogToggleStatus, handleCatalogDelete, handleCatalogReview, sendEmptyPublicList, applyCatalogStatusFilter, CATALOG_PUBLIC_VISIBILITY_QUERY, deriveApprovalStatus } from './shared';
 import { toOptionalString, toStringArray } from './inputCoercion';
 import { modelCreateSchema, modelUpdateSchema, rejectionSchema } from '@esparex/core/validators/catalog.validator';
 import CategoryQueryBuilder from '@esparex/core/utils/CategoryQueryBuilder';
-import { MAX_MODEL_TREE_DEPTH } from '@esparex/core/services/catalog/CatalogHierarchyService';
-import { updateModelHierarchyTransactionally } from '@esparex/core/services/catalog/CatalogHierarchyService';
+import { MAX_MODEL_TREE_DEPTH, updateModelHierarchyTransactionally } from '@esparex/core/domains/catalog/application/services/CatalogHierarchyService';
 import { getCache } from '@esparex/core/utils/redisCache';
 import { catalogCacheKey, applyCacheWriteThrough, normalizeOptionalObjectIdQuery, normalizeBooleanQuery, populateModelVariants, applyModelHierarchyPayload, logModelDuplicateCandidates } from './adminCatalogShared';
