+++ backend/api/src/controllers/catalogRequestController.ts
@@ -1,6 +1,6 @@
 import type { Request, Response } from 'express';
 import { type ICatalogRequest } from '@esparex/core/models/CatalogRequest';
-import * as CatalogRequestService from '@esparex/core/services/catalog/CatalogRequestService';
+import * as CatalogRequestService from '@esparex/core/domains/catalog/application/services/CatalogRequestService';
 import { sendPaginatedResponse, sendSuccessResponse } from '../utils/respond';
 import { sendErrorResponse } from '../utils/errorResponse';
 import { AppError } from '@esparex/core/utils/AppError';
