--- core/src/services/catalog/CatalogBrandModelService.ts
@@ -1,1 +1,0 @@
-export * from "../../domains/catalog/application/services/CatalogBrandModelService";
