--- core/src/services/catalog/CatalogCategoryService.ts
@@ -1,1 +1,0 @@
-export * from "../../domains/catalog/application/services/CatalogCategoryService";
