--- core/src/services/catalog/CatalogHierarchyService.ts
@@ -1,1 +1,0 @@
-export * from "../../domains/catalog/application/services/CatalogHierarchyService";
