--- core/src/services/catalog/CatalogImportService.ts
@@ -1,1 +1,0 @@
-export * from "../../domains/catalog/application/services/CatalogImportService";
