--- core/src/services/catalog/CatalogNotificationService.ts
@@ -1,1 +1,0 @@
-export * from "../../domains/catalog/application/services/CatalogNotificationService";
