--- core/src/services/catalog/CatalogOrchestrator.ts
@@ -1,2 +1,0 @@
-export * from "../../domains/catalog/application/services/CatalogOrchestrator";
-export { default } from "../../domains/catalog/application/services/CatalogOrchestrator";
