--- core/src/services/catalog/CatalogReferenceService.ts
@@ -1,1 +1,0 @@
-export * from "../../domains/catalog/application/services/CatalogReferenceService";
