--- core/src/services/catalog/CatalogRequestService.ts
@@ -1,1 +1,0 @@
-export * from "../../domains/catalog/application/services/CatalogRequestService";
