--- core/src/services/catalog/CatalogResolutionPolicy.ts
@@ -1,1 +1,0 @@
-export * from "../../domains/catalog/application/services/CatalogResolutionPolicy";
