--- core/src/services/catalog/CatalogSearchGovernanceService.ts
@@ -1,1 +1,0 @@
-export * from "../../domains/catalog/application/services/CatalogSearchGovernanceService";
