--- core/src/services/catalog/CatalogSparePartService.ts
@@ -1,1 +1,0 @@
-export * from "../../domains/catalog/application/services/CatalogSparePartService";
