--- core/src/services/catalog/CatalogValidationService.ts
@@ -1,1 +1,0 @@
-export * from "../../domains/catalog/application/services/CatalogValidationService";
